+++ config/initializers/trusted_proxies.rb
@@ -7,8 +7,7 @@ Rails.application.configure do
     config.conjur_disable_trusted_proxies_cache
 
   Rack::Request.ip_filter = Conjur::TrustedProxyFilter.new(
-    options: {
-      disable_cache: disable_cache
-    }
+    config: Rails.application.config.conjur_config,
+    disable_cache: disable_cache
   )
 end