+++ bin/conjur-cli.rb
@@ -3,6 +3,7 @@
 require 'gli'
 require 'net/http'
 require 'uri'
+require 'open3'
 
 include GLI::App
 
@@ -81,7 +82,10 @@ command :server do |c|
 
     if account
       if options["password-from-stdin"]
-        password = stdin_input
+        # Rake is interpreting raw commas in the password as
+        # delimiting addtional arguments to rake itself. 
+        # Reference: https://github.com/ruby/rake/blob/a842fb2c30cc3ca80803fba903006b1324a62e9a/lib/rake/application.rb#L163
+        password = stdin_input.gsub(',', '\,')
         system("rake 'account:create_with_password[#{account},#{password}]'")\
           or exit $?.exitstatus
       else
@@ -233,7 +237,10 @@ $ conjurctl account create [--password-from-stdin] --name myorg
       connect
 
       if options["password-from-stdin"]
-        password = stdin_input
+        # Rake is interpreting raw commas in the password as
+        # delimiting addtional arguments to rake itself. 
+        # Reference: https://github.com/ruby/rake/blob/a842fb2c30cc3ca80803fba903006b1324a62e9a/lib/rake/application.rb#L163
+        password = stdin_input.gsub(',', '\,')
         exec("rake 'account:create_with_password[#{account},#{password}]'")
       else
         exec("rake 'account:create[#{account}]'")
@@ -274,7 +281,16 @@ command :role do |cgrp|
       connect
 
       fail 'key retrieval failed' unless args.map { |id|
-        system("rake 'role:retrieve-key[#{id}]'")
+        stdout, stderr, = Open3.capture3("rake 'role:retrieve-key[#{id}]'")
+
+        if stderr.empty?
+          # Only print last line of stdout to omit server config logging
+          puts(stdout.split("\n").last)
+          true
+        else
+          $stderr.puts(stderr)
+          false
+        end
       }.all?
     end
   end