+++ config/application.rb
@@ -21,6 +21,9 @@ Bundler.require(*Rails.groups)
 $LOAD_PATH.push(File.expand_path("../../engines/conjur_audit/lib", __FILE__))
 require 'conjur_audit'
 
+# Must require because lib folder hasn't been loaded yet
+require './lib/conjur/conjur_config'
+
 module Conjur
   class Application < Rails::Application
     # Settings in config/environments/* take precedence over those specified here.
@@ -56,6 +59,25 @@ module Conjur
     # checks are sufficient to verify the usage of an environment variable.
     ENV.each_pair do |(k, v)|
       ENV[k] = nil if v =~ /^\s*$/ # is all whitespace
+    end
+
+    # Allows us to use a config file that doesn't group values by Rails env.
+    config.anyway_config.future.unwrap_known_environments = true
+
+    config.anyway_config.default_config_path = "/etc/conjur/config"
+
+    # Create a single instance of the ConjurConfig object for this process that
+    # loads configuration on server startup. This prevents config values from
+    # being loaded fresh every time a ConjurConfig object is instantiated, which
+    # could lead to inconsistent behavior.
+    #
+    # We create this in application.rb instead of an initializer so that it's
+    # guaranteed to be available for other initializers to use.
+    config.conjur_config = Conjur::ConjurConfig.new
+
+    puts "Loaded configuration:"
+    config.conjur_config.attribute_sources.each do |k,v|
+      puts "- #{k} from #{v}"
     end
   end
 end